requirements.txt
fastapi
uvicorn
pillow
tensorflow
scipy
python-multipart